fix: polish quit snapshot dialog
