+++ Sources/AppDelegate.swift
@@ -547,158 +547,6 @@ final class CmuxMainThreadTurnProfiler {
 #endif
 
 @MainActor
-private final class QuitSessionSnapshotProgressController: NSObject {
-    private let panel: NSPanel
-    private let titleLabel: NSTextField
-    private let detailLabel: NSTextField
-    private let progressIndicator: NSProgressIndicator
-    private let skipButton: NSButton
-    private var didMarkSkipping = false
-
-    private(set) var shouldSkipRemainingScrollback = false
-
-    init(totalWorkspaces: Int) {
-        titleLabel = NSTextField(
-            labelWithString: String(localized: "dialog.quitSessionSnapshot.title", defaultValue: "Saving session")
-        )
-        detailLabel = NSTextField(
-            labelWithString: String(
-                localized: "dialog.quitSessionSnapshot.message",
-                defaultValue: "Saving terminal scrollback before quitting."
-            )
-        )
-        progressIndicator = NSProgressIndicator()
-        skipButton = NSButton(
-            title: String(
-                localized: "dialog.quitSessionSnapshot.skipRemaining",
-                defaultValue: "Skip Remaining Scrollback"
-            ),
-            target: nil,
-            action: nil
-        )
-        panel = NSPanel(
-            contentRect: NSRect(x: 0, y: 0, width: 420, height: 158),
-            styleMask: [.titled],
-            backing: .buffered,
-            defer: false
-        )
-
-        super.init()
-
-        panel.identifier = NSUserInterfaceItemIdentifier("cmux.quitSessionSnapshotProgress")
-        panel.title = titleLabel.stringValue
-        panel.isReleasedWhenClosed = false
-        panel.hidesOnDeactivate = false
-        panel.level = .modalPanel
-        panel.collectionBehavior = [.canJoinAllSpaces, .fullScreenAuxiliary]
-
-        titleLabel.font = .boldSystemFont(ofSize: 15)
-        detailLabel.font = .systemFont(ofSize: 13)
-        detailLabel.textColor = .secondaryLabelColor
-        detailLabel.lineBreakMode = .byWordWrapping
-        detailLabel.maximumNumberOfLines = 2
-
-        progressIndicator.style = .bar
-        progressIndicator.minValue = 0
-        progressIndicator.maxValue = Double(max(totalWorkspaces, 1))
-        progressIndicator.doubleValue = 0
-        progressIndicator.isIndeterminate = totalWorkspaces == 0
-
-        skipButton.target = self
-        skipButton.action = #selector(skipRemainingScrollback(_:))
-
-        let contentView = NSView()
-        panel.contentView = contentView
-        [titleLabel, detailLabel, progressIndicator, skipButton].forEach {
-            $0.translatesAutoresizingMaskIntoConstraints = false
-            contentView.addSubview($0)
-        }
-
-        NSLayoutConstraint.activate([
-            titleLabel.topAnchor.constraint(equalTo: contentView.topAnchor, constant: 18),
-            titleLabel.leadingAnchor.constraint(equalTo: contentView.leadingAnchor, constant: 20),
-            titleLabel.trailingAnchor.constraint(equalTo: contentView.trailingAnchor, constant: -20),
-
-            detailLabel.topAnchor.constraint(equalTo: titleLabel.bottomAnchor, constant: 6),
-            detailLabel.leadingAnchor.constraint(equalTo: titleLabel.leadingAnchor),
-            detailLabel.trailingAnchor.constraint(equalTo: titleLabel.trailingAnchor),
-
-            progressIndicator.topAnchor.constraint(equalTo: detailLabel.bottomAnchor, constant: 16),
-            progressIndicator.leadingAnchor.constraint(equalTo: titleLabel.leadingAnchor),
-            progressIndicator.trailingAnchor.constraint(equalTo: titleLabel.trailingAnchor),
-
-            skipButton.topAnchor.constraint(equalTo: progressIndicator.bottomAnchor, constant: 14),
-            skipButton.trailingAnchor.constraint(equalTo: titleLabel.trailingAnchor),
-            skipButton.bottomAnchor.constraint(lessThanOrEqualTo: contentView.bottomAnchor, constant: -16),
-        ])
-    }
-
-    func show(relativeTo window: NSWindow?) {
-        if let window {
-            panel.center()
-            let windowFrame = window.frame
-            var panelFrame = panel.frame
-            panelFrame.origin.x = windowFrame.midX - panelFrame.width / 2
-            panelFrame.origin.y = windowFrame.midY - panelFrame.height / 2
-            panel.setFrame(panelFrame, display: false)
-        } else {
-            panel.center()
-        }
-        if progressIndicator.isIndeterminate {
-            progressIndicator.startAnimation(nil)
-        }
-        panel.orderFrontRegardless()
-        panel.displayIfNeeded()
-    }
-
-    func update(currentIndex: Int, total: Int) {
-        guard !shouldSkipRemainingScrollback else {
-            markSkipping()
-            return
-        }
-        let resolvedTotal = max(total, 1)
-        let resolvedCurrent = min(max(currentIndex, 0), resolvedTotal)
-        progressIndicator.isIndeterminate = false
-        progressIndicator.maxValue = Double(resolvedTotal)
-        progressIndicator.doubleValue = Double(resolvedCurrent)
-        let format = String(
-            localized: "dialog.quitSessionSnapshot.progressFormat",
-            defaultValue: "Saving scrollback %d of %d..."
-        )
-        detailLabel.stringValue = String.localizedStringWithFormat(format, resolvedCurrent, resolvedTotal)
-        panel.displayIfNeeded()
-    }
-
-    func markSkipping() {
-        guard !didMarkSkipping else { return }
-        didMarkSkipping = true
-        progressIndicator.isIndeterminate = true
-        progressIndicator.startAnimation(nil)
-        detailLabel.stringValue = String(
-            localized: "dialog.quitSessionSnapshot.skippingRemaining",
-            defaultValue: "Skipping remaining scrollback..."
-        )
-        panel.displayIfNeeded()
-    }
-
-    func close() {
-        progressIndicator.stopAnimation(nil)
-        panel.close()
-    }
-
-    @objc private func skipRemainingScrollback(_ sender: NSButton) {
-        shouldSkipRemainingScrollback = true
-        sender.isEnabled = false
-        markSkipping()
-    }
-}
-
-@MainActor
-private final class QuitSessionSnapshotProgressState {
-    var visitedWorkspaceCount = 0
-}
-
-@MainActor
 final class AppDelegate: NSObject, NSApplicationDelegate, UNUserNotificationCenterDelegate, NSMenuItemValidation {
     nonisolated(unsafe) static var shared: AppDelegate?
